MainForm.h - przeczyszczone main.cpp - lekko wyczyszczone, std::thread zamiast wątków winapi poziom ostrzeżeń dla projektu podniesiony z W3 do W4 wszystkie ostrzeżenia traktowane jako błędy printowanie wskaźników w konsoli - zmieniłem rzutowanie z DWORD do void* w MainForm.h większość funkcji nie używa wszystkich parametrów co generuje ostrzeżenie C4100 - wyłączyłem je w tym pliku usunąłem całkiem buildy x64 bo biorąc pod uwagę działanie programu adresy działały bardzo różnie
zmieniony kod działa poprawnie (testowany)
forma poprawnie pokazuje się w designerze po zmianach - w nawiązaniach do komentarza (vs 19)

+++ DevTale/NosString.cpp
@@ -22,8 +22,8 @@ NosString::NosString(char* i8_string)
 
 	this->i8_string_ = static_cast<char*>(malloc(this->i32_length_ + 8 + 1));
 
-	*reinterpret_cast<unsigned long*>(this->i8_string_ + 0x00) = 1;
-	*reinterpret_cast<unsigned long*>(this->i8_string_ + 0x04) = this->i32_length_;
+	*reinterpret_cast<size_t*>(this->i8_string_ + 0x00) = 1;
+	*reinterpret_cast<size_t*>(this->i8_string_ + 0x04) = this->i32_length_;
 
 	memcpy(this->i8_string_ + 0x08, i8_string, this->i32_length_);
 
@@ -35,7 +35,7 @@ char* NosString::get() const
 	return this->i8_string_ + 0x08;
 }
 
-unsigned long NosString::length() const
+size_t NosString::length() const
 {
 	return this->i32_length_;
 }
